refactor(chat): extract projection schemas into their own module
The per-tool projection schemas move out of ref-field-map.ts (1596 -> 421
lines) into projections.ts, keeping their precise inferred builder types
so consumers can read v.InferInput over the real entries; the map widens
at its RefMediationEntry boundary as before, so its surface is unchanged.
Adds AssertNoExtraFields for payloads that are not object literals.
